+++ internal/dockerhost/client_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/api/types/registry"
 	"github.com/docker/go-connections/nat"
-	. "github.com/fluxplane/fluxplane-dex/plugins/docker"
 )
 
 func TestContainerCreateConfigMapsDockerOptions(t *testing.T) {