+++ internal/dockerhost/docker_e2e_test.go
@@ -7,8 +7,6 @@ import (
 	"strings"
 	"testing"
 	"time"
-
-	. "github.com/fluxplane/fluxplane-dex/plugins/docker"
 )
 
 func TestDockerE2EContainerImageNetworkVolumeAndCopy(t *testing.T) {